Remove unneeded extra complexity on command process
