--- src/main/kotlin/com/artifactgames/copyplash/model/GameSettings.kt
@@ -1,3 +1,0 @@
-package com.artifactgames.copyplash.model
-
-data class GameSettings(val locale:String)
