--- src/main/kotlin/com/artifactgames/copyplash/model/Question.kt
@@ -1,4 +1,0 @@
-package com.artifactgames.copyplash.model
-
-
-data class Question(val id: String, val description: String)
