+++ src/main/kotlin/com/artifactgames/copyplash/model/RequestModels.kt
@@ -5,3 +5,5 @@ import com.artifactgames.copyplash.type.States
 
 
 data class GameMode(val gameMode: GameModes, val state: States)
+
+data class GameSettings(val locale:String)
